+++ bot/consumers.py
@@ -1,26 +1,9 @@
-import functools
-import json
-
-from channels.generic.websocket import AsyncWebsocketConsumer
+from channels.generic.websocket import AsyncJsonWebsocketConsumer
 
 from .models import DiscordUser
 
 
-def proccess_json():
-    """
-    Gets `text_data` and transforms it into a :class:`dict` for sugar syntax.
-    """
-
-    def wrapper(func):
-        @functools.wraps(func)
-        async def wrapped(*args, **kwargs):
-            kwargs['text_data'] = json.loads(kwargs['text_data'])
-            return await func(*args, **kwargs)
-        return wrapped
-    return wrapper
-
-
-class BotConnectionOnRegisterConsumer(AsyncWebsocketConsumer):
+class BotConnectionOnRegisterConsumer(AsyncJsonWebsocketConsumer):
     """
     WebSocket for Bot Actions.
     """
@@ -28,15 +11,14 @@ class BotConnectionOnRegisterConsumer(AsyncWebsocketConsumer):
     async def connect(self):
         await super(BotConnectionOnRegisterConsumer, self).connect()
 
-    @proccess_json()
-    async def receive(self, text_data=None, bytes_data=None):
-        discord_id = text_data['discord_id']
+    async def receive_json(self, content, **kwargs):
+        discord_id = content['discord_id']
         data = {
             'exists': False
         }
         if DiscordUser.objects.filter(pk=discord_id).exists():
             # Discord User exists
             data['exists'] = True
-            await self.send(json.dumps(data))
+            await self.send_json(data)
         else:
-            await self.send(json.dumps(data))
+            await self.send_json(data)
